requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
openai==1.2.4
httpx==0.25.0
python-multipart==0.0.6